Reorders lib and fixes linting issues
